feat: 抽奖中置规则过滤

+++ alias-market-domain/src/main/java/com/alias/domain/strategy/model/entity/RuleMatterEntity.java
@@ -9,7 +9,7 @@ public class RuleMatterEntity {
 
     private Long strategyId;
 
-    private Integer awardId;
+    private Long awardId;
 
     private String ruleModel;
 }
